devtools/common/cli/devtool/pyproject.toml
[project]
name = "dt"
version = "0.1.0"
description = "Unified CLI for development tools"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.21.1",
    "pyyaml>=6.0",
]

[project.scripts]
dt = "dt.core.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dt"]
